+++ src/CDA.PCML/Program.cs
@@ -35,12 +35,12 @@ namespace CDA.PCML
             PrepareOutputFolder(folderPath);
             
             var pcmlCda = new PCMLSample();
-            pcmlCda.MaxPopulatedPCMLAuthorHealthcareProviderSample_1A("PCML_AuthorHealthcareProvider_1A_Max.xml");
-            pcmlCda.MinPopulatedPCMLAuthorHealthcareProviderSample_1A("PCML_AuthorHealthcareProvider_1A_Min.xml");
+            pcmlCda.MaxPopulatedPCMLAuthorHealthcareProviderSample_1A("PSML_AuthorHealthcareProvider_1A_Max.xml");
+            pcmlCda.MinPopulatedPCMLAuthorHealthcareProviderSample_1A("PSML_AuthorHealthcareProvider_1A_Min.xml");
 
             //Test Custom Narrative - not currently allowed
-            //pcmlCda.MaxPopulatedPCMLAuthorHealthcareProviderSample_1B("PCML_AuthorHealthcareProvider_1B_Max.xml");
-            //pcmlCda.MinPopulatedPCMLAuthorHealthcareProviderSample_1B("PCML_AuthorHealthcareProvider_1B_Min.xml");
+            //pcmlCda.MaxPopulatedPCMLAuthorHealthcareProviderSample_1B("PSML_AuthorHealthcareProvider_1B_Max.xml");
+            //pcmlCda.MinPopulatedPCMLAuthorHealthcareProviderSample_1B("PSML_AuthorHealthcareProvider_1B_Min.xml");
 
         }